requirements.txt
groq==0.11.0
httpx>=0.27.0
psycopg2-binary==2.9.9
APScheduler==3.10.4
python-dotenv==1.0.1
pytz==2024.1